refactor(mcp): split session tool definitions
